reolink_api/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.9" and python_version < "4.0"
anyio==4.9.0 ; python_version >= "3.9" and python_version < "4.0"
certifi==2025.6.15 ; python_version >= "3.9" and python_version < "4.0"
charset-normalizer==3.4.2 ; python_version >= "3.9" and python_version < "4.0"
click==8.1.8 ; python_version == "3.9"
click==8.2.1 ; python_version >= "3.10" and python_version < "4.0"
colorama==0.4.6 ; python_version >= "3.9" and python_version < "4.0" and (platform_system == "Windows" or sys_platform == "win32")
exceptiongroup==1.3.0 ; python_version >= "3.9" and python_version < "3.11"
fastapi==0.110.3 ; python_version >= "3.9" and python_version < "4.0"
h11==0.16.0 ; python_version >= "3.9" and python_version < "4.0"
httptools==0.6.4 ; python_version >= "3.9" and python_version < "4.0"
idna==3.10 ; python_version >= "3.9" and python_version < "4.0"
numpy==1.26.4 ; python_version >= "3.9" and python_version < "4.0"
opencv-python-headless==4.11.0.86 ; python_version >= "3.9" and python_version < "4.0"
pillow==10.4.0 ; python_version >= "3.9" and python_version < "4.0"
pydantic-core==2.33.2 ; python_version >= "3.9" and python_version < "4.0"
pydantic==2.11.7 ; python_version >= "3.9" and python_version < "4.0"
python-dotenv==1.1.1 ; python_version >= "3.9" and python_version < "4.0"
pyyaml==6.0.2 ; python_version >= "3.9" and python_version < "4.0"
requests==2.32.4 ; python_version >= "3.9" and python_version < "4.0"
sniffio==1.3.1 ; python_version >= "3.9" and python_version < "4.0"
starlette==0.37.2 ; python_version >= "3.9" and python_version < "4.0"
typing-extensions==4.14.0 ; python_version >= "3.9" and python_version < "4.0"
typing-inspection==0.4.1 ; python_version >= "3.9" and python_version < "4.0"
urllib3==2.5.0 ; python_version >= "3.9" and python_version < "4.0"
uvicorn==0.30.6 ; python_version >= "3.9" and python_version < "4.0"
uvloop==0.21.0 ; python_version >= "3.9" and python_version < "4.0" and sys_platform != "win32" and sys_platform != "cygwin" and platform_python_implementation != "PyPy"
watchfiles==1.1.0 ; python_version >= "3.9" and python_version < "4.0"
websockets==15.0.1 ; python_version >= "3.9" and python_version < "4.0"
